Added Diffster service registration and usage sample
Updated `DiffsterExtensions.cs` to include a new method `AddDiffster<T, TOutput>` for registering a `Diffster` service with a specified formatter. This method validated null arguments and added a singleton instance to the service collection. The `using` directive for `Microsoft.Extensions.DependencyInjection` was also added.

In `Program.cs`, a new usage sample was added to demonstrate formatting differences as key-value pairs. A `keyValuePairFormatter` function was defined to convert `PropertyDifference` objects into `KeyValuePair<string, (object FirstValue, object SecondValue)>`, and the formatted differences were printed to the console.